+++ csatar-plugins/csatar/models/PermissionBasedAccess.php
@@ -118,6 +118,9 @@ class PermissionBasedAccess extends Model
 
     public static function getTranslatedAttributeNames(string $organizationTypeModelName = null): array
     {
+        if($organizationTypeModelName == parent::getModelName()){
+            return [];
+        }
         if ((is_array(self::$translatedAttributeNames) && !array_key_exists($organizationTypeModelName, self::$translatedAttributeNames)) ||
             !is_array(self::$translatedAttributeNames)
         ) {